🔧 WeatherSimple 모델 디렉토리 분리
